+++ gulpfile.js
@@ -1,5 +1,4 @@
 var gulp = require('gulp');
-var less = require('gulp-less');
 var livereload = require('gulp-livereload');
 var minifycss = require('gulp-minify-css');
 var rename = require('gulp-rename');
@@ -8,30 +7,12 @@ var sass = require('gulp-sass');
 var adminFolder = 'nope/lib/';
 var themeFolder = 'nope/theme/default/';
 
-gulp.task('less', function() {
-  gulp.src(adminFolder + 'assets/less/app.less')
-    .pipe(less())
-    .pipe(minifycss())
-    .pipe(rename('app.min.css'))
-    .pipe(gulp.dest(adminFolder + 'assets/css'))
-    .pipe(livereload());
-});
-
 gulp.task('sass', function() {
   gulp.src(adminFolder + 'assets/scss/app.scss')
     .pipe(sass())
     .pipe(minifycss())
     .pipe(rename('app.min.css'))
     .pipe(gulp.dest(adminFolder + 'assets/css'))
-    .pipe(livereload());
-});
-
-gulp.task('theme-less', function() {
-  gulp.src(themeFolder + 'assets/less/style.less')
-    .pipe(less())
-    .pipe(minifycss())
-    .pipe(rename('style.min.css'))
-    .pipe(gulp.dest(themeFolder + 'assets/css'))
     .pipe(livereload());
 });
 
@@ -43,23 +24,14 @@ gulp.task('else', function(cb) {
 gulp.task('watch', function() {
   livereload.listen();
   gulp.watch([
-    adminFolder + 'assets/less/*.less',
-    adminFolder + 'assets/less/**/*.less'
-  ], ['less']);
-  gulp.watch([
     adminFolder + 'assets/scss/*.scss',
     adminFolder + 'assets/scss/**/*.scss'
   ], ['sass']);
   gulp.watch([
-    themeFolder + 'assets/less/*.less',
-    themeFolder + 'assets/less/**/*.less'
-  ], ['theme-less']);
-  gulp.watch([
     'index.php',
     'nope/*.*',
     'nope/**/*.*',
-    '!nope/**/*.less',
-    '!nope/**/*.sass',
+    '!nope/**/*.scss',
     '!nope/**/*.css',
     '!nope/lib/vendor/**',
     '!nope/storage/**/*.*'