+++ nope/lib/views/modal/content.php
@@ -3,7 +3,7 @@
     <iframe ng-src="{{url}}"></iframe>
   </nope-modal-body>
   <nope-modal-footer>
-    <a href="" class="btn btn-default" nope-modal-close>Close</a>
-    <a href="" class="btn btn-success" ng-click="$parent.selection.length?$parent.onSelect(selection):false;" ng-disabled="!$parent.selection.length">Select <span ng-show="$parent.selection.length">({{selection.length}} item<span ng-show="$parent.selection.length>1">s</span>)</span></a>
+    <a href="" class="btn btn-secondary btn-lg" nope-modal-close>Close</a>
+    <a href="" class="btn btn-success btn-lg" ng-click="$parent.selection.length?$parent.onSelect(selection):false;" ng-disabled="!$parent.selection.length">Select <span ng-show="$parent.selection.length">({{selection.length}} item<span ng-show="$parent.selection.length>1">s</span>)</span></a>
   </nope-modal-footer>
 </nope-modal>