+++ nope/lib/views/modal/upload.php
@@ -1,19 +1,25 @@
 <nope-modal class="modal--upload" title="Upload media">
   <nope-modal-body>
-    <a href="" nope-upload="onDone();" accept="{{$parent.accept}}" on-progress="progressList" status="uploadingStatus" class="btn btn-block btn-outline-secondary btn-upload" ng-show="!uploadingStatus"><i class="fa fa-upload"></i>Choose files to upload</a>
-    <ul class="list-group" ng-if="progressList">
-      <li class="list-group-item" ng-class="{'list-group-item-success':(value.percentage===100),'list-group-item-danger':value.error}" ng-repeat="(key, value) in progressList">
-        <div class="row">
-          <div class="col-8">{{key}} {{value.errorMessage}}</div>
-          <div class="col-2">
-            <progress class="progress" ng-class="{'progress-info':(value.percentage<100 && !value.error),'progress-success':(value.percentage===100), 'progress-danger':value.error}" value="{{value.percentage}}" max="100"></progress>
-          </div>
+    <div ng-show="!progressList.status">
+      <a href="" nope-upload="onDone();" accept="{{$parent.accept}}" on-progress="progressList" class="btn btn-block btn-outline-secondary btn-upload" ng-show="!progressList.status"><i class="fa fa-upload"></i>Choose files to upload</a>
+    </div>
+    <ul class="list-group" ng-if="progressList.status">
+      <li class="list-group-item" ng-class="{'list-group-item-success':(value.percentage===100),'list-group-item-danger':value.error}" ng-repeat="(key, value) in progressList.list">
+        <div class="d-flex w-100 justify-content-between">
+          <h6>{{key}}</h6>
+          <small ng-if="value.percentage && value.percentage<100">{{value.percentage}}%</small>
+          <i class="fa fa-check-circle" ng-if="value.percentage===100"></i>
+          <i class="fa fa-exclamation-circle" ng-if="value.errorMessage"></i>
         </div>
+        <p class="text-danger" ng-show="value.errorMessage"><small>{{value.errorMessage}}</small></p>
       </li>
     </ul>
-    <div ng-show="!uploadingStatus">
+    <div ng-if="!progressList.status">
       <hr />
-      <nope-import on-done="onDoneImport();"></nope-import>
+      <nope-import on-done="onDoneFooter();"></nope-import>
     </div>
   </nope-modal-body>
+  <nope-modal-footer ng-show="showFooter">
+    <a href="" class="btn btn-success btn-lg" nope-modal-close>Done</a>
+  </nope-modal-footer>
 </nope-modal>